+++ super_editor_quill/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.1.0-dev.7
+* Feature: Customizable block-level embed parsers.
+
+## 0.1.0-dev.6
+* Adjustment: A custom `Editor` can now be supplied when parsing deltas.
+
+## 0.1.0-dev.5
+* Feature: Customizable inline embed parsers and serializers.
+
+## 0.1.0-dev.4
+* Fix: Avoid throwing an exception when using ambiguous custom delta formats.
+
 ## 0.1.0-dev.3
 * Switch `super_editor` dependency from GitHub version to pre-release version `0.3.0-dev.3`.